+++ src/App.tsx
@@ -7,8 +7,8 @@ import Login from 'pages/Auth/Login';
 import Home from 'pages/Home';
 import Post from 'pages/Post';
 import Search from 'pages/Search';
-import SearchPage from 'pages/Auth/SearchIDPassword';
-import DistinctionNumber from 'pages/Auth/SearchIDPassword/page/DistinctionNumber';
+import FindIdPassword from 'pages/Auth/SearchIDPassword';
+import VerifyField from 'pages/Auth/SearchIDPassword/page/VerifyField';
 import SearchQueryItemList from 'pages/Search/components/SearchQueryItemList';
 import { Routes, Route } from 'react-router-dom';
 import ChangePassword from 'pages/Auth/SearchIDPassword/page/ChangePassword';
@@ -24,10 +24,10 @@ function App(): JSX.Element {
         <Route path="/post" element={<Post />} />
         <Route path="/login" element={<Login />} />
         <Route path="/search" element={<Search />} />
-        <Route path="/search-id" element={<SearchPage search="id" />} />
-        <Route path="/search-pw" element={<SearchPage search="password" />} />
-        <Route path="/search-id/number" element={<DistinctionNumber />} />
-        <Route path="/search-pw/change" element={<ChangePassword />} />
+        <Route path="/find-id" element={<FindIdPassword find="id" />} />
+        <Route path="/find-password" element={<FindIdPassword find="password" />} />
+        <Route path="/find/verify" element={<VerifyField />} />
+        <Route path="/find-password/change" element={<ChangePassword />} />
         <Route path="/search/:searchQuery" element={<SearchQueryItemList />} />
       </Routes>
       <Toast />
