+++ src/pages/Auth/SearchIDPassword/page/VerifyCode.module.scss
@@ -12,9 +12,15 @@ input[type="number"]::-webkit-inner-spin-button {
 .form {
   display: flex;
   justify-content: center;
+  flex-direction: column;
 }
 
 .form__input {
+  display: flex;
+  justify-content: center;
+}
+
+.form__input--block {
   width: 50px;
   height: 50px;
   background-color: #eeeeee;
